back/requirements.txt
asgiref==3.4.1
botocore==1.21.56
Django==3.2.8
django-storages==1.12
djangorestframework==3.12.4
jmespath==0.10.0
mysqlclient==2.0.3
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
